Pin generic enum constructor arity diagnostic schema
