+++ tests/docs_truth/public_docs.rs
@@ -294,6 +294,7 @@ fn diagnostics_catalog_documents_json_stable_codes() {
         "tests/fixtures/ir_json/diagnostics_duplicate_generic_impl.golden.json",
         "tests/fixtures/ir_json/diagnostics_generic_function_arity.golden.json",
         "tests/fixtures/ir_json/diagnostics_generic_struct_constructor_arity.golden.json",
+        "tests/fixtures/ir_json/diagnostics_generic_enum_constructor_arity.golden.json",
         "tests/fixtures/ir_json/diagnostics_generic_result_method_arity.golden.json",
         "tests/fixtures/ir_json/diagnostics_generic_requires_arity.golden.json",
         "tests/fixtures/ir_json/diagnostics_generic_impl_arity.golden.json",
