+++ tests/fixtures/ir_json/diagnostics_generic_enum_constructor_arity.golden.json
@@ -1,0 +1,29 @@
+{
+  "format": "zen.diagnostics.v0",
+  "semantic_status": "diagnostic",
+  "files": [
+    {
+      "id": 0,
+      "path": "$TMP/generic_enum_constructor_arity.zen"
+    }
+  ],
+  "diagnostics": [
+    {
+      "severity": "error",
+      "code": "E5001",
+      "message": "generic enum `Option` expects 1 type arguments, found 2",
+      "span": {
+        "file_id": 0,
+        "path": "$TMP/generic_enum_constructor_arity.zen",
+        "start": 63,
+        "end": 96,
+        "line": 7,
+        "column": 13
+      },
+      "labels": [],
+      "notes": [],
+      "context": [],
+      "suggested_fixes": []
+    }
+  ]
+}
