+++ tests/integration/cli_build/frontend_json.rs
@@ -3,6 +3,8 @@ use std::process::Command;
 
 #[path = "frontend_json/diagnostics_behavior_association_golden.rs"]
 mod diagnostics_behavior_association_golden;
+#[path = "frontend_json/diagnostics_generic_arity_golden.rs"]
+mod diagnostics_generic_arity_golden;
 #[path = "frontend_json/diagnostics_golden.rs"]
 mod diagnostics_golden;
 #[path = "frontend_json/diagnostics_json.rs"]
