+++ tests/integration/cli_build/frontend_json/diagnostics_golden.rs
@@ -210,109 +210,6 @@ Point.implements(PrettyJson) {
 }
 
 #[test]
-fn emit_json_diagnostics_generic_function_arity_schema_matches_golden() {
-    let tmp = tempfile::tempdir().expect("create temp dir");
-    let zen_path = tmp.path().join("generic_function_arity.zen");
-    std::fs::write(
-        &zen_path,
-        r#"
-identity<T> = (value: T) T {
-    value
-}
-
-main = () i32 {
-    identity<i32, StaticString>(1)
-}
-"#,
-    )
-    .expect("write generic function arity source");
-
-    let output = Command::new(env!("CARGO_BIN_EXE_zen"))
-        .args(["emit-json", "diagnostics", zen_path.to_str().unwrap()])
-        .output()
-        .expect("run zen emit-json diagnostics");
-
-    assert!(
-        !output.status.success(),
-        "zen emit-json diagnostics should fail on generic function arity: stdout={}, stderr={}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    let actual = String::from_utf8(output.stdout).expect("diagnostics stdout is UTF-8");
-    let json: serde_json::Value =
-        serde_json::from_str(&actual).expect("diagnostics stdout is JSON");
-    assert_eq!(
-        json["diagnostics"]
-            .as_array()
-            .expect("diagnostics array")
-            .len(),
-        1,
-        "generic function arity diagnostics should not emit inference or argument followups: {json}"
-    );
-
-    let normalized = actual.replace(tmp.path().to_str().expect("tmp path is UTF-8"), "$TMP");
-    let expected_path =
-        fixture("tests/fixtures/ir_json/diagnostics_generic_function_arity.golden.json");
-    let expected = std::fs::read_to_string(&expected_path)
-        .unwrap_or_else(|err| panic!("read {}: {err}", expected_path.display()));
-
-    assert_eq!(normalized.trim(), expected.trim());
-}
-
-#[test]
-fn emit_json_diagnostics_generic_struct_constructor_arity_schema_matches_golden() {
-    let tmp = tempfile::tempdir().expect("create temp dir");
-    let zen_path = tmp.path().join("generic_struct_constructor_arity.zen");
-    std::fs::write(
-        &zen_path,
-        r#"
-Box<T>: {
-    value: T
-}
-
-main = () i32 {
-    boxed = Box<i32, StaticString> { value: 1 }
-    boxed.value
-}
-"#,
-    )
-    .expect("write generic struct constructor arity source");
-
-    let output = Command::new(env!("CARGO_BIN_EXE_zen"))
-        .args(["emit-json", "diagnostics", zen_path.to_str().unwrap()])
-        .output()
-        .expect("run zen emit-json diagnostics");
-
-    assert!(
-        !output.status.success(),
-        "zen emit-json diagnostics should fail on generic struct constructor arity: stdout={}, stderr={}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    let actual = String::from_utf8(output.stdout).expect("diagnostics stdout is UTF-8");
-    let json: serde_json::Value =
-        serde_json::from_str(&actual).expect("diagnostics stdout is JSON");
-    assert_eq!(
-        json["diagnostics"]
-            .as_array()
-            .expect("diagnostics array")
-            .len(),
-        1,
-        "generic struct constructor arity diagnostics should not emit field mismatch followups: {json}"
-    );
-
-    let normalized = actual.replace(tmp.path().to_str().expect("tmp path is UTF-8"), "$TMP");
-    let expected_path =
-        fixture("tests/fixtures/ir_json/diagnostics_generic_struct_constructor_arity.golden.json");
-    let expected = std::fs::read_to_string(&expected_path)
-        .unwrap_or_else(|err| panic!("read {}: {err}", expected_path.display()));
-
-    assert_eq!(normalized.trim(), expected.trim());
-}
-
-#[test]
 fn emit_json_diagnostics_generic_result_method_arity_schema_matches_golden() {
     let tmp = tempfile::tempdir().expect("create temp dir");
     let zen_path = tmp.path().join("generic_result_method_arity.zen");
